Feat/ci integration (#24)
* feat: add CI integration

* feat: add main contract test

* feat: add travis CI integration and code coverage reports

* :new: :alembic: Add CircleCI config for NodeJS/npm (#23)

This commit removes the Travis CI config and replaces it with a CircleCI
config file. The CircleCI pipeline has one job, `npm-test`, which runs
all the commands that were used in the Travis config file.

Additionally, this CircleCI config also uses an Orb for NodeJS. This
automatically takes care of package caching between builds, so your
builds launch and run quickly if dependencies do not change.

I also left a comment with the Coveralls integration, which supports
test outputs in Mocha (i.e. what Truffle is built on top of for
JavaScript). You can configure it and get it working once the tests are
fixed and passing.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

Co-authored-by: Justin W. Flory <jflory7@gmail.com>